Change csv output format
Since we now use a slightly different calculation to fit the decode phase model,
we add a few more parameters in the csv.
Here is a brief definition for these parameters:
 - prompt_tokens: The total number of tokens sent to the model.
 - prompt_eval_time: The duration between the time we send the prompt and the
   time we receive the first token from the model.
 - tokens_per_sec: The token rate measured during the decode phase.
 - decode_start_token: The token counts at the time when we start the
   measurement of decode time.
 - decode_time: The duration of the decode phase.
In this commit, we added different handling mechanism to gather the above
parameters for benchmarking against ollama and maas.